+++ src/bus.rs
@@ -46,8 +46,8 @@ pub use super::driver::Speed;
 /// ```no_run
 /// use imxrt_usbd::BusAdapter;
 ///
-/// # struct Ps; use imxrt_usbd::Instance as Inst;
-/// # unsafe impl imxrt_usbd::Peripherals for Ps { fn instance(&self) -> Inst { panic!() } }
+/// # struct Ps;
+/// # unsafe impl imxrt_usbd::Peripherals for Ps { fn usb(&self) -> *const () { panic!() } fn usbphy(&self) -> *const () { panic!() }}
 /// static EP_MEMORY: imxrt_usbd::EndpointMemory<1024> = imxrt_usbd::EndpointMemory::new();
 /// static EP_STATE: imxrt_usbd::EndpointState = imxrt_usbd::EndpointState::max_endpoints();
 ///