+++ src/lib.rs
@@ -44,8 +44,6 @@ pub use state::{EndpointState, MAX_ENDPOINTS};
 ///
 /// - USB core registers
 /// - USB PHY registers
-/// - USB non-core registers
-/// - USB analog registers
 ///
 /// When an instance of `Peripherals` exists, you must make sure that
 /// nothing else, other than the owner of the `Peripherals` object,
@@ -58,10 +56,9 @@ pub use state::{EndpointState, MAX_ENDPOINTS};
 /// operation. Incorrect usage, or failure to ensure exclusive
 /// ownership, could lead to data races and incorrect USB functionality.
 ///
-/// By implementing this trait, you ensure that the [`Instance`]
-/// identifier is valid for your chip. Not all i.MX RT peripherals
-/// have a USB2 peripheral instance, so you must ensure that the
-/// implementation is correct for your chip.
+/// All pointers are expected to point at the starting register block
+/// for the specified peripheral. Calls to the functions must return the
+/// the same value every time they're called.
 ///
 /// # Example
 ///
@@ -75,18 +72,19 @@ pub use state::{EndpointState, MAX_ENDPOINTS};
 ///
 /// ```no_run
 /// # mod ral {
-/// #   use core::ops::Deref; pub struct Instance; impl Deref for Instance { type Target = u32; fn deref(&self) -> &u32 { unsafe { &*(0x402e0200 as *const u32)} } }
+/// #   pub struct RegisterBlock;
+/// #   use core::ops::Deref; pub struct Instance; impl Deref for Instance { type Target = RegisterBlock; fn deref(&self) -> &RegisterBlock { unsafe { &*(0x402e0200 as *const RegisterBlock)} } }
 /// #   pub fn take() -> Result<Instance, ()> { Ok(Instance) }
-/// #   pub mod usb { pub use super::Instance; pub mod USB1 { pub use super::super::take; } }
-/// #   pub mod usbphy { pub use super::Instance; pub mod USBPHY1 { pub use super::super::take; } }
+/// #   pub mod usb { pub use super::{Instance, RegisterBlock}; pub mod USB1 { pub use super::super::take; } }
+/// #   pub mod usbphy { pub use super::{Instance, RegisterBlock}; pub mod USBPHY1 { pub use super::super::take; } }
 /// #   pub mod usbnc { pub use super::Instance; pub mod USBNC1 { pub use super::super::take; } }
 /// #   pub mod usb_analog { pub use super::Instance; pub mod USB_ANALOG { pub use super::super::take; } }
 /// # }
 /// use ral::usb;
 ///
 /// struct Peripherals {
-///     _usb: ral::usb::Instance,
-///     _phy: ral::usbphy::Instance,
+///     usb: ral::usb::Instance,
+///     phy: ral::usbphy::Instance,
 ///     _nc: ral::usbnc::Instance,
 ///     _analog: ral::usb_analog::Instance,
 /// }
@@ -95,8 +93,8 @@ pub use state::{EndpointState, MAX_ENDPOINTS};
 ///     /// Panics if the instances are already taken
 ///     fn usb1() -> Peripherals {
 ///         Self {
-///             _usb: ral::usb::USB1::take().unwrap(),
-///             _phy: ral::usbphy::USBPHY1::take().unwrap(),
+///             usb: ral::usb::USB1::take().unwrap(),
+///             phy: ral::usbphy::USBPHY1::take().unwrap(),
 ///             _nc: ral::usbnc::USBNC1::take().unwrap(),
 ///             _analog: ral::usb_analog::USB_ANALOG::take().unwrap(),
 ///         }
@@ -108,8 +106,13 @@ pub use state::{EndpointState, MAX_ENDPOINTS};
 /// // guaranteed to be singletons. Given this approach, no one else
 /// // can safely access the USB registers.
 /// unsafe impl imxrt_usbd::Peripherals for Peripherals {
-///     fn instance(&self) -> imxrt_usbd::Instance {
-///         imxrt_usbd::Instance::USB1
+///     fn usb(&self) -> *const () {
+///         let rb: &ral::usb::RegisterBlock = &self.usb;
+///         (rb as *const ral::usb::RegisterBlock).cast()
+///     }
+///     fn usbphy(&self) -> *const () {
+///         let rb: &ral::usbphy::RegisterBlock = &self.phy;
+///         (rb as *const ral::usbphy::RegisterBlock).cast()
 ///     }
 /// }
 ///
@@ -122,26 +125,8 @@ pub use state::{EndpointState, MAX_ENDPOINTS};
 /// );
 /// ```
 pub unsafe trait Peripherals {
-    /// Returns the instance identifier for the core registers
-    ///
-    /// **Warning**: some i.MX RT peripherals have only one USB peripheral,
-    /// USB1. The behavior is undefined if you return `Instance::USB2` on
-    /// one of these systems.
-    fn instance(&self) -> Instance;
-}
-
-/// USB instance identifiers
-///
-/// These are *not* USB standards or speeds. They indicate if this
-/// is the USB1 register instance, or the USB2 register instance.
-///
-/// Note that some i.MX RT processors only have one USB instance (USB1).
-/// On those systems, it is invalid to ever indicate the USB2 instance.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[allow(clippy::upper_case_acronyms)] // USB naming makes sense here
-pub enum Instance {
-    /// The first USB register instance
-    USB1,
-    /// The second USB register instance
-    USB2,
+    /// Returns the pointer to the USB register block.
+    fn usb(&self) -> *const ();
+    /// Returns the pointer to the USBPHY register block.
+    fn usbphy(&self) -> *const ();
 }